backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0

# AI and API clients
openai==1.3.0
httpx==0.25.2

# File handling
aiofiles==23.2.1

# WebSocket support
websockets==12.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Caching (optional)
redis==5.0.1

# Image processing
Pillow==10.0.1

# Data analysis and visualization
matplotlib==3.8.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Web scraping and parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration and templating
PyYAML==6.0.1
Jinja2==3.1.2

# CLI tools
click==8.1.7
rich==13.7.0
typer==0.9.0

# Validation and utilities
validators==0.22.0
python-magic==0.4.27
chardet==5.2.0
pathspec==0.11.2

# Code quality and formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2

# Security scanning
bandit==1.7.5
safety==2.3.4

# Pre-commit hooks
pre-commit==3.5.0

# Performance and monitoring
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Additional utilities
python-slugify==8.0.1
shortuuid==1.0.11
filelock==3.13.1
psutil==5.9.6
schedule==1.2.0

# Development dependencies (optional)
jupyter==1.0.0
ipython==8.17.2

# Additional AI/ML libraries (optional)
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

